refactor #51: 상수 분리 및 코드 정리
- 길이가 긴 정적 값을 contants로 분리
- util 함수 분리
