+++ fe/src/components/IssueList/Issue.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 
 import Profile from '@common/Profile';
 import Label from '@common/Label';
-import { LabelRow, elapseTime } from './IssueList';
+import { LabelRow } from './IssueList';
+import { ElapseTime } from '@utils/getTimeElapsed';
 import { ReactComponent as AlertCircle } from '@assets/alertCircle.svg';
 import { ReactComponent as Archive } from '@assets/archive.svg';
 import { ReactComponent as Milestone } from '@assets/milestone.svg';
@@ -14,7 +15,7 @@ interface Props {
   profileUrl: string;
   isOpen: boolean;
   labelList: LabelRow[];
-  elapseTime: elapseTime;
+  elapseTime: ElapseTime;
   milestoneName?: string;
   onIssueTitleClick: (id: number) => void;
 }
