+++ fe/src/constants/MainPage.ts
@@ -1,0 +1,24 @@
+export const FILTER_DROPDOWN_LIST = [
+  {
+    id: 0,
+    title: '열린 이슈',
+  },
+  {
+    id: 1,
+    title: '내가 작성한 이슈',
+  },
+  {
+    id: 2,
+    title: '나에게 할당된 이슈',
+  },
+  {
+    id: 3,
+    title: '내가 댓글을 남긴 이슈',
+  },
+  {
+    id: 4,
+    title: '닫힌 이슈',
+  },
+];
+
+export const API_URL = 'http://43.200.199.205:8080/api/';
